Start a replacement for Handle based on sans I/O
The new Client structure implements the netlink protocol as a state machine for
handling requests and responses.  Notably, it never stores a socket or performs
any reads or writes, requiring the callee to perform their own I/O.  These
changes are non-breaking yet, but the Handle struct will eventually be removed
in favor of this style.

https://sans-io.readthedocs.io/